+++ Adyen/ApiSerialization/IMessagePayload.cs
@@ -21,10 +21,7 @@
 //  */
 #endregion
 
-namespace Adyen.CloudApiSerialization
+namespace Adyen.ApiSerialization
 {
-    internal interface IMessagePayloadSerializer<out T> where T : IMessagePayload
-    {
-        IMessagePayload Deserialize(string messagePayloadJson);
-    }
+    public interface IMessagePayload { }
 }
