+++ Adyen/ApiSerialization/MessagePayloadSerializer.cs
@@ -20,32 +20,15 @@
 //  * See the LICENSE file for more info.
 //  */
 #endregion
-using Newtonsoft.Json;
-using System;
 
-namespace Adyen.CloudApiSerialization.Converter
+namespace Adyen.ApiSerialization
 {
-    internal class JsonBase64Converter : JsonConverter
+    internal class MessagePayloadSerializer<T> : IMessagePayloadSerializer<T> where T : IMessagePayload
     {
-        public override bool CanConvert(Type objectType)
+        public IMessagePayload Deserialize(string messagePayloadJson)
         {
-            return true;
+            return Converter.JSonConvertDeserializerWrapper<T>.DeserializeObject(messagePayloadJson);
         }
 
-        public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
-        {
-            return System.Text.Encoding.UTF8.GetString((Convert.FromBase64String((string)reader.Value)));
-        }
-
-        public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
-        {
-            var json = JsonConvert.SerializeObject(value, Formatting.None,
-                            new JsonSerializerSettings
-                            {
-                                NullValueHandling = NullValueHandling.Ignore
-                            });
-
-            writer.WriteValue(Convert.ToBase64String(System.Text.Encoding.UTF8.GetBytes(json)));
-        }
     }
 }
