+++ Adyen/Model/Nexo/SaleData.cs
@@ -21,7 +21,7 @@
 //  */
 #endregion
 
-using Adyen.CloudApiSerialization.Converter;
+using Adyen.ApiSerialization.Converter;
 using Adyen.Model.Terminal;
 using Newtonsoft.Json;
